Correciones aplicadas
