+++ FitAdmin/src/main/java/org/example/Interfaces/AccionBusqueda.java
@@ -1,0 +1,8 @@
+package org.example.Interfaces;
+
+import org.example.GUI.JfrAgregarNuevoCliente;
+import org.example.Modelo.Cliente;
+
+public interface AccionBusqueda {
+    void ejecutar(String busqueda,JfrAgregarNuevoCliente nuevoCliente, Cliente cliente);
+}
